add support for non-qwerty layout
The support is done via a class that consist of a mapping from KC to
others.

Such mapping are similar to what is done in QMK therefore a script is
provided to bootstrap their definition.

Existing doc on how to add internationla keycode is removed as those
keycode are now all integrated in the base KC definition.