[doc:rgb] fix link to neopixel lib
The link is hardcoded to a commit from 2018-12-21 (version 3.3.5) while
we are currently at 6.3.3.

The rgb makes use of the fact that NeoPixel is a subclass of PixelBuf
but that was not true back then. Hence some people following the link in
the doc get a neopixel library incompatible with kmk.