Shrink ./kmk/ with `nox` & `mpy-cross`.
Shrink ./kmk/ with `nox` & `mpy-cross`.

The nice!nano CircuitPython doesn't leave enough storage for an
uncompiled ./kmk/ folder, with main.py and Bluetooth libraries.

This introduces a soft dependancy to kmk on `nox`, as a OS-agnostic
alternative to `make`. Bonus is the use of native python to describe
steps / options / sessions.