made analogio scanner (not mux yet)
Made AnalogIo scanner that can scan single input.
muxing should just be adding scanning with shifing the pins on the mux so not to hard but will need to take in muxers as an array in themself so multiple can be supported at once

this is mostly baised on the Digital Io scanner but I see no exsamples as to where it gets it's offset from.